resolve warnings
